+++ src/audit/mod.rs
@@ -1,3 +1,9 @@
+mod builder;
+mod record;
+
+pub use builder::{AuditRecordBuilder, AuditRecordMetadata};
+pub use record::{AuditEventType, AuditRecord, AuditRecordDetails, AuditStatus};
+
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct AuditReference {
     id: String,